4_Return_predictions/requirements.txt
numpy==1.26.4
pandas==2.2.2
PyJWT==2.10.1
boto3==1.35.81
awslambdaric==3.0.0
